Fixes #989 WARN: require cycle
