+++ lib/ProgressComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import TrackPlayer from './index.js';
+import TrackPlayer from './trackPlayer.js';
 
 class ProgressComponent extends Component {
