+++ lib/hooks.js
@@ -1,5 +1,5 @@
 import { useEffect, useState, useDebugValue, useRef } from 'react';
-import * as TrackPlayer from './index';
+import * as TrackPlayer from './trackPlayer';
 import TrackPlayerEvents from './eventTypes';
 
 /**
